+++ scilab/routines/scicos/matmul_i16n.c
@@ -30,8 +30,8 @@ void matmul_i16n(scicos_block *block,int flag)
 		    D=D + C;}
 		    t=D-(int)(D/(k))*((k));
 		    if ((t>=k/2)|(-(t)>=k/2))
-		       {if (t>=0) (t)=(-((k/2))+abs(t-((int)((t)/((k/2))))*((k/2))));
-			else (t)=-(-((k/2))+abs(t-((int)((t)/((k/2))))*((k/2))));}
+		       {if (t>=0) (t)=(-((k/2))+abs(t-((double)((int)((t)/((k/2)))))*((k/2))));
+			else (t)=-(-((k/2))+abs(t-((double)((int)((t)/((k/2)))))*((k/2))));}
 		    y[jl]=(short)t;
 		  }
 	     }
